+++ R/generating_initialpop.R
@@ -26,7 +26,7 @@
 
 generating_initialpop <- function (
         nstudies,
-        nsamples = 5000,
+        nsamples = 3000,
         action_mode,
         par_names, lower_bounds, upper_bounds,
         uncertainty = c("par_level", "func_level", "diffusion"),