+++ R/loss.R
@@ -26,7 +26,7 @@
 loss <- function (
         par, data,
         nsims = NULL,
-        nsamples = 5000,
+        nsamples = 3000,
         exec_threshold = 1,
         imag_threshold = 0.5,
         uncertainty = c("par_level", "func_level", "diffusion"),